+++ Changelog.md
@@ -1,6 +1,7 @@
 -----------Changelog for v1.1.3-----------
 
 - Added config to control whether other players should receive join/leave messages when a player (un-)vanishes
+- Added configs to change the messages the player receives when vanishing/unvanishing
 - Fixed that Command Blocks are not able to execute /v by default
 
 -----------Changelog for v1.1.2-----------